+++ CHANGES.rst
@@ -5,11 +5,12 @@ Changelog
 1.16 (unreleased)
 -----------------
 
-- Skip setting a default page if the default is already set to the same value. 
+- Skip setting a default page if the default is already set to the same value.
   [mamico]
 - Fixed data extraction of Archetypes ATFilefields to concatenate all `Pdata` chunks, when the FileField data was still stored in Filestorage and not not migrated to use blobstorage (using plone.app.blobs).
   [@jnptk]
 - Bugfix to prevent clearing existing relations from catalog when migrating into sites with existing content [ThibautBorn]
+- Enable option to not export the portal settings (local roles, default pages, portlets) [ThibautBorn]
 
 1.15 (2025-07-15)
 -----------------